Add self-handing for the run function

+++ ranhNenLamToolProject/scheduledtask.cpp
@@ -9,19 +9,28 @@ ScheduledTask::ScheduledTask(QTime scheduledTime, QString fileLocation, Schedule
 
 void ScheduledTask::run()
 {
+    //If the task is active
     if (mScheduleState == ScheduledTask::STATE_ACTIVE)
     {
-        QString executableName = mFileLocation.mid(mFileLocation.lastIndexOf("\\") + 1);
-        if (mScheduledAction == ScheduledTask::ACTION_CLOSE)
+        //And its time has come
+        QTime currentTime = QTime::currentTime();
+        if (currentTime.hour() == mScheduledTime.hour()
+            &&  currentTime.minute() == mScheduledTime.minute()
+            &&  currentTime.second() == mScheduledTime.second())
         {
-            QProcess::startDetached("taskkill /im " + executableName + " /f");
+            QString executableName = mFileLocation.mid(mFileLocation.lastIndexOf("\\") + 1);
+            // Perform the scheduled action
+            if (mScheduledAction == ScheduledTask::ACTION_CLOSE)
+            {
+                QProcess::startDetached("taskkill /im " + executableName + " /f");
+            }
+            else if (mScheduledAction == ScheduledTask::ACTION_OPEN)
+            {
+                QProcess::startDetached(mFileLocation);
+            }
+            //Auto deactivate itself
+            mScheduleState = ScheduledTask::STATE_INACTIVE;
         }
-        else if (mScheduledAction == ScheduledTask::ACTION_OPEN)
-        {
-            QProcess::startDetached(mFileLocation);
-        }
-        //Auto deactivate itself
-
     }
 }
